Request from Brian_Camire@URSCorp.com I've spoken to Michael about this issue, and we concluded that:
1. Users might access the application from different computers that have ArcMap installed in different folders,

2. KDOT wants to minimize the administrative burden, and,

3. It is unlikely that KDOT's security policies would prevent the end user from being able to read the registry keys we would need to find out where ArcMap was installed.


As a result, KDOT would like to use the "look the path up in the registry" approach unless you have more compelling reasons not to. The cleanest way of implementing this might be for you add an argument to the GET_X_CMD functions in your X_KANSAS_EPFS package for us to pass you the path to the ArcMap executable. What do you think?

KDOT would like us to resolve this issue as soon as possible.

So added new parameter pi_arcmap_exe to get procedures, if this is null when passed it will use the value in the product option.



git-svn-id: svn://127.0.0.1/customer/trunk@57 37afec09-bd54-7940-9fac-cf987e4a02e6